# review-service/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /helloworld/{name}:
        get:
            tags:
                - Greeter
            description: Sends a greeting
            operationId: Greeter_SayHello
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.HelloReply'
    /v1/appeal/audit:
        post:
            tags:
                - Review
            description: O端评价申诉审核
            operationId: Review_AuditAppeal
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.AuditAppealRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.AuditAppealReply'
    /v1/review:
        post:
            tags:
                - Review
            description: C端创建评价
            operationId: Review_CreateReview
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.CreateReviewRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.CreateReviewReply'
    /v1/review/appeal:
        post:
            tags:
                - Review
            description: B端申诉评价
            operationId: Review_AppealReview
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.AppealReviewRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.AppealReviewReply'
    /v1/review/audit:
        post:
            tags:
                - Review
            description: O端审核评价
            operationId: Review_AuditReview
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.AuditReviewRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.AuditReviewReply'
    /v1/review/reply:
        post:
            tags:
                - Review
            description: B端回复评价
            operationId: Review_ReplyReview
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.ReplyReviewRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.ReplyReviewReply'
    /v1/review/{reviewID}:
        get:
            tags:
                - Review
            description: C端获取评价详情
            operationId: Review_GetReview
            parameters:
                - name: reviewID
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.GetReviewReply'
    /v1/{userID}/reviews:
        get:
            tags:
                - Review
            description: C端查看userID下所有评价
            operationId: Review_ListReviewByUserID
            parameters:
                - name: userID
                  in: path
                  required: true
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: size
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.ListReviewByUserIDReply'
components:
    schemas:
        api.review.v1.AppealReviewReply:
            type: object
            properties:
                appealID:
                    type: string
            description: 对评价进行申诉的返回值
        api.review.v1.AppealReviewRequest:
            type: object
            properties:
                reviewID:
                    type: string
                storeID:
                    type: string
                reason:
                    type: string
                content:
                    type: string
                picInfo:
                    type: string
                videoInfo:
                    type: string
            description: AppealReviewRequest 申诉评价的请求参数
        api.review.v1.AuditAppealReply:
            type: object
            properties: {}
            description: 对申诉进行审核的返回值
        api.review.v1.AuditAppealRequest:
            type: object
            properties:
                appealID:
                    type: string
                reviewID:
                    type: string
                status:
                    type: integer
                    format: int32
                opUser:
                    type: string
                opRemarks:
                    type: string
            description: 对申诉进行审核的请求
        api.review.v1.AuditReviewReply:
            type: object
            properties:
                reviewID:
                    type: string
                status:
                    type: integer
                    format: int32
            description: 审核评价的返回值
        api.review.v1.AuditReviewRequest:
            type: object
            properties:
                reviewID:
                    type: string
                status:
                    type: integer
                    format: int32
                opUser:
                    type: string
                opReason:
                    type: string
                opRemarks:
                    type: string
            description: 审核评价的请求
        api.review.v1.CreateReviewReply:
            type: object
            properties:
                reviewID:
                    type: string
            description: 创建评价的回复
        api.review.v1.CreateReviewRequest:
            type: object
            properties:
                userID:
                    type: string
                orderID:
                    type: string
                storeID:
                    type: string
                score:
                    type: integer
                    format: int32
                serviceScore:
                    type: integer
                    format: int32
                expressScore:
                    type: integer
                    format: int32
                content:
                    type: string
                picInfo:
                    type: string
                videoInfo:
                    type: string
                anonymous:
                    type: boolean
            description: 创建评价的参数
        api.review.v1.GetReviewReply:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/api.review.v1.ReviewInfo'
            description: 获取评价详情的响应
        api.review.v1.ListReviewByUserIDReply:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.review.v1.ReviewInfo'
            description: 用户评价列表的返回值
        api.review.v1.ReplyReviewReply:
            type: object
            properties:
                replyID:
                    type: string
            description: 回复评价的返回值
        api.review.v1.ReplyReviewRequest:
            type: object
            properties:
                reviewID:
                    type: string
                storeID:
                    type: string
                content:
                    type: string
                picInfo:
                    type: string
                videoInfo:
                    type: string
            description: 回复评价的请求
        api.review.v1.ReviewInfo:
            type: object
            properties:
                reviewID:
                    type: string
                userID:
                    type: string
                orderID:
                    type: string
                score:
                    type: integer
                    format: int32
                serviceScore:
                    type: integer
                    format: int32
                expressScore:
                    type: integer
                    format: int32
                content:
                    type: string
                picInfo:
                    type: string
                videoInfo:
                    type: string
                status:
                    type: integer
                    format: int32
            description: 评价信息
        helloworld.v1.HelloReply:
            type: object
            properties:
                message:
                    type: string
            description: The response message containing the greetings
tags:
    - name: Greeter
      description: The greeting service definition.
    - name: Review
      description: 定义评价服务